change np.float64 to floaat

+++ src/foraging_gui/MyFunctions.py
@@ -732,7 +732,7 @@ class GenerateTrials():
         reward_refills = np.vstack([p_Ls >= random_number_L, p_Rs >= random_number_R])
         optimal_choices = np.argmax([p_Ls, p_Rs], axis=0)  # Greedy choice, assuming the agent knows the groundtruth
         optimal_rewards = reward_refills[0][optimal_choices==0].sum() + reward_refills[1][optimal_choices==1].sum()
-        for_eff_optimal_random_seed = reward_rate / (optimal_rewards / len(optimal_choices))
+        for_eff_optimal_random_seed = float(reward_rate / (optimal_rewards / len(optimal_choices)))
         
         return for_eff_optimal, for_eff_optimal_random_seed
 
@@ -783,7 +783,7 @@ class GenerateTrials():
                 reward_remain[this_choice[t]] = 0
             
             if reward_optimal_random_seed:                
-                for_eff_optimal_random_seed = reward_rate / (reward_optimal_random_seed / len(p_Ls))
+                for_eff_optimal_random_seed = float(reward_rate / (reward_optimal_random_seed / len(p_Ls)))
             else:
                 for_eff_optimal_random_seed = np.nan
